Edited init function to make unit testing easier / faster.
When running unit tests you don't always want to be hitting external
resources.  Like downloading the rss from vice, mainly because that can
be slow and also because its hard to write deterministic tests against
test data that is always changing.

So to avoid having to hit the external URL the first step is to make
the URL paramaeter in the init optional.

One way to do this would have been:

```python
def __init__(self, url=None):
... rest of code here...
```

However this will "show" developers that they don't need to pass in the
url parameter.  Because we generally want developers to pass in the URL
parameter when they are using this library (but we don't want to pass
it in when we are running our tests).  I am leaving the url parameter
as required, but I put in a check to skip the construction if the url
is None.  i.e.

```python
def __init__(self, url):
    if url:
        ... init code here ...
```

This gives us a "backdoor" to make this function easier to use when
testing.  And we will do that in the next commit.