Changed the unit test to not required an external resource.
For this change I did a few things.

1.  The RSSViceTester class has been changed to not read from a url but
rather load a file (which is a snapshot of an rss feed).  And then
setup the RSSVice class as if it had read from the URL.  This is done
in the `setUp` function.

2.  In order to get the vice.rss file I just used the following code:

```python
self.vice = ViceRSS("http://www.vice.com/rss")
with open("vice.rss", "w") as f:
    f.write(self.vice.raw)
```

that creates the file and then we can just use that file from this
point forward for our testing.

From here you can no expand your testsing to do things like.  Ensuring
you get the specific word could, that article_split, returns the
correct number of articles.  Etc etc... try to get the tests updated
and offering more coverage, then we can look at the actual class and
see what we can do to improve upon that.